+++ examples/simple.rs
@@ -1,8 +1,7 @@
 extern crate libtest_mimic;
 
+use libtest_mimic::{Arguments, Failed, Trial};
 use std::{thread, time};
-use libtest_mimic::{Arguments, Trial, Failed};
-
 
 fn main() {
     let args = Arguments::from_args();
@@ -17,7 +16,6 @@ fn main() {
 
     libtest_mimic::run(&args, tests).exit();
 }
-
 
 // Tests