+++ tests/common/mod.rs
@@ -1,8 +1,7 @@
-use std::{path::Path, iter::repeat_with, collections::HashMap};
 use pretty_assertions::assert_eq;
+use std::{iter::repeat_with, path::Path};
 
 use libtest_mimic::{run, Arguments, Conclusion, Trial};
-
 
 const TEMPDIR: &str = env!("CARGO_TARGET_TMPDIR");
 
@@ -14,23 +13,24 @@ pub fn args<const N: usize>(args: [&str; N]) -> Arguments {
 
 pub fn do_run(mut args: Arguments, tests: Vec<Trial>) -> (Conclusion, String) {
     // Create path to temporary file.
-    let suffix = repeat_with(fastrand::alphanumeric).take(10).collect::<String>();
+    let suffix = repeat_with(fastrand::alphanumeric)
+        .take(10)
+        .collect::<String>();
     let path = Path::new(&TEMPDIR).join(format!("libtest_mimic_output_{suffix}.txt"));
 
     args.logfile = Some(path.display().to_string());
 
     let c = run(&args, tests);
-    let output = std::fs::read_to_string(&path)
-        .expect("Can't read temporary logfile");
-    std::fs::remove_file(&path)
-        .expect("Can't remove temporary logfile");
+    let output = std::fs::read_to_string(&path).expect("Can't read temporary logfile");
+    std::fs::remove_file(&path).expect("Can't remove temporary logfile");
     (c, output)
 }
 
 /// Removes shared indentation so that at least one line has no indentation
 /// (no leading spaces).
 pub fn clean_expected_log(s: &str) -> String {
-    let shared_indent = s.lines()
+    let shared_indent = s
+        .lines()
         .filter(|l| l.contains(|c| c != ' '))
         .map(|l| l.bytes().take_while(|b| *b == b' ').count())
         .min()
@@ -52,21 +52,30 @@ pub fn clean_expected_log(s: &str) -> String {
 
 /// Best effort tool to check certain things about a log that might have all
 /// tests randomly ordered.
+#[cfg(feature = "multithreaded")]
 pub fn assert_reordered_log(actual: &str, num: u64, expected_lines: &[&str], tail: &str) {
     let actual = actual.trim();
     let (first_line, rest) = actual.split_once('\n').expect("log has too few lines");
     let (middle, last_line) = rest.rsplit_once('\n').expect("log has too few lines");
 
-
-    assert_eq!(first_line, &format!("running {} test{}", num, if num == 1 { "" } else { "s" }));
+    assert_eq!(
+        first_line,
+        &format!("running {} test{}", num, if num == 1 { "" } else { "s" })
+    );
     assert!(last_line.contains(tail));
+
+    use std::collections::HashMap;
 
     let mut actual_lines = HashMap::new();
     for line in middle.lines().map(|l| l.trim()).filter(|l| !l.is_empty()) {
         *actual_lines.entry(line).or_insert(0) += 1;
     }
 
-    for expected in expected_lines.iter().map(|l| l.trim()).filter(|l| !l.is_empty()) {
+    for expected in expected_lines
+        .iter()
+        .map(|l| l.trim())
+        .filter(|l| !l.is_empty())
+    {
         match actual_lines.get_mut(expected) {
             None | Some(0) => panic!("expected line \"{expected}\" not in log"),
             Some(num) => *num -= 1,
@@ -97,19 +106,27 @@ macro_rules! assert_log {
             }
         }
 
+        if let Some(pos) = actual.rfind("\"exec_time\":") {
+            actual.truncate(pos);
+            actual.push_str("\"exec_time\": 0.000000000 }");
+        }
+
         assert_eq!(actual, expected);
     };
 }
 
-pub fn check(
-    mut args: Arguments,
-    mut tests: impl FnMut() -> Vec<Trial>,
-    num_running_tests: u64,
+pub fn check<'a>(
+    #[allow(unused_mut)] mut args: Arguments,
+    mut tests: impl FnMut() -> Vec<Trial<'a>>,
+    #[allow(unused_variables)] num_running_tests: u64,
     expected_conclusion: Conclusion,
     expected_output: &str,
 ) {
     // Run in single threaded mode
-    args.test_threads = Some(1);
+    #[cfg(feature = "multithreaded")]
+    {
+        args.test_threads = Some(1);
+    }
     let (c, out) = do_run(args.clone(), tests());
     let expected = crate::common::clean_expected_log(expected_output);
     let actual = {
@@ -119,19 +136,29 @@ pub fn check(
     assert_eq!(actual.trim(), expected.trim());
     assert_eq!(c, expected_conclusion);
 
-    // Run in multithreaded mode.
-    let (c, out) = do_run(args, tests());
-    assert_reordered_log(
-        &out,
-        num_running_tests,
-        &expected_output.lines().collect::<Vec<_>>(),
-        &conclusion_to_output(&c),
-    );
-    assert_eq!(c, expected_conclusion);
+    #[cfg(feature = "multithreaded")]
+    {
+        // Run in multithreaded mode.
+        let (c, out) = do_run(args, tests());
+        assert_reordered_log(
+            &out,
+            num_running_tests,
+            &expected_output.lines().collect::<Vec<_>>(),
+            &conclusion_to_output(&c),
+        );
+        assert_eq!(c, expected_conclusion);
+    }
 }
 
+#[cfg(feature = "multithreaded")]
 fn conclusion_to_output(c: &Conclusion) -> String {
-    let Conclusion { num_filtered_out, num_passed, num_failed, num_ignored, num_measured } = *c;
+    let Conclusion {
+        num_filtered_out,
+        num_passed,
+        num_failed,
+        num_ignored,
+        num_measured,
+    } = *c;
     format!(
         "test result: {}. {} passed; {} failed; {} ignored; {} measured; {} filtered out;",
         if num_failed > 0 { "FAILED" } else { "ok" },